requirements.txt
streamlit
stegano
pillow
cryptography